+++ lib/src/extensions/tile_set.dart
@@ -5,13 +5,13 @@ import "package:tmx_parser/tmx_parser.dart";
 extension TileSetExtensions on TileSet {
   /// Creates an [Offset] from [tileOffset]
   Vector2 get offset => Vector2(
-        this.tileOffset.x,
-        this.tileOffset.y,
+        this.tileOffset.x.toDouble(),
+        this.tileOffset.y.toDouble(),
       );
 
   Vector2 get tileSize => Vector2(
-        tileWidth,
-        tileHeight,
+        tileWidth.toDouble(),
+        tileHeight.toDouble(),
       );
 
   /// Calculates the [Offset] needed to extract the correct tile from the [TileSet]
@@ -20,8 +20,8 @@ extension TileSetExtensions on TileSet {
     int tileY = tileIndex ~/ this.columns;
 
     Vector2 sourceOffset = Vector2(
-      tileX * this.tileWidth,
-      tileY * this.tileHeight,
+      tileX * this.tileWidth.toDouble(),
+      tileY * this.tileHeight.toDouble(),
     );
 
     return sourceOffset;
