+++ lib/src/extensions/tmx_map.dart
@@ -1,9 +1,9 @@
-import 'package:flame/extensions.dart';
-import 'package:tmx_parser/tmx_parser.dart';
+import "package:flame/extensions.dart";
+import "package:tmx_parser/tmx_parser.dart";
 
 extension TmxMapExtensions on TmxMap {
   Vector2 get tileSize => Vector2(
-        tileWidth,
-        tileHeight,
+        tileWidth.toDouble(),
+        tileHeight.toDouble(),
       );
 }
